Move all advertisers' client nodes to a dedicated servername, to support `origin` header
